requirements.txt
aniso8601==6.0.0
Click==7.0
Flask==1.0.2
Flask-JWT-Extended==3.18.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
jws==0.1.3
MarkupSafe==1.1.1
passlib==1.7.1
psycopg2-binary==2.8.2
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.3
Werkzeug==0.15.2
